[utils] Reduce compile time dependencies for CJobManager, CJobQueue, CJob, IJobCallback.
