+++ xbmc/GUILargeTextureManager.h
@@ -11,6 +11,7 @@
 #include "guilib/AspectRatio.h"
 #include "guilib/TextureManager.h"
 #include "threads/CriticalSection.h"
+#include "utils/IJobCallback.h"
 #include "utils/Job.h"
 
 #include <memory>
