+++ xbmc/cores/paplayer/PAPlayer.h
@@ -14,7 +14,7 @@
 #include "cores/IPlayer.h"
 #include "threads/CriticalSection.h"
 #include "threads/Thread.h"
-#include "utils/Job.h"
+#include "utils/IJobCallback.h"
 
 #include <atomic>
 #include <list>
