+++ xbmc/music/MusicLibraryQueue.h
@@ -10,7 +10,7 @@
 
 #include "settings/LibExportSettings.h"
 #include "threads/CriticalSection.h"
-#include "utils/JobManager.h"
+#include "utils/JobQueue.h"
 
 #include <map>
 #include <set>
