+++ xbmc/platform/win10/storage/Win10StorageProvider.h
@@ -9,7 +9,6 @@
 #pragma once
 
 #include "storage/IStorageProvider.h"
-#include "utils/Job.h"
 
 #include <atomic>
 #include <string>
