+++ xbmc/utils/CMakeLists.txt
@@ -37,6 +37,7 @@ set(SOURCES ActorProtocol.cpp
             HttpResponse.cpp
             InfoLoader.cpp
             JobManager.cpp
+            JobQueue.cpp
             JSONVariantParser.cpp
             JSONVariantWriter.cpp
             LabelFormatter.cpp
@@ -124,6 +125,7 @@ set(HEADERS ActorProtocol.h
             HttpResponse.h
             IArchivable.h
             IBufferObject.h
+            IJobCallback.h
             ILocalizer.h
             InfoLoader.h
             IPlatformLog.h
@@ -134,9 +136,11 @@ set(HEADERS ActorProtocol.h
             IXmlDeserializable.h
             Job.h
             JobManager.h
+            JobQueue.h
             JSONVariantParser.h
             JSONVariantWriter.h
             LabelFormatter.h
+            LambdaJob.h
             LangCodeExpander.h
             LegacyPathTranslation.h
             Locale.h
