+++ xbmc/utils/RecentlyAddedJob.h
@@ -8,7 +8,7 @@
 
 #pragma once
 
-#include "Job.h"
+#include "utils/Job.h"
 
 enum ERecentlyAddedFlag
 {
