+++ xbmc/video/dialogs/GUIDialogSubtitles.h
@@ -10,7 +10,7 @@
 
 #include "guilib/GUIDialog.h"
 #include "threads/CriticalSection.h"
-#include "utils/JobManager.h"
+#include "utils/JobQueue.h"
 
 #include <string>
