+++ xbmc/weather/WeatherJob.h
@@ -9,6 +9,7 @@
 #pragma once
 
 #include "WeatherManager.h"
+#include "utils/Job.h"
 
 #include <map>
 #include <string>
